ADDED Dynamic Content rendering for Modals
