+++ src/components/MartianModalContent.js
@@ -1,0 +1,46 @@
+import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+
+import { Modal, Button } from "@material-ui/core";
+
+function rand() {
+  return Math.round(Math.random() * 20) - 10;
+}
+
+function getModalStyle() {
+  const top = 50 + rand();
+  const left = 50 + rand();
+
+  return {
+    top: `${top}%`,
+    left: `${left}%`,
+    transform: `translate(-${top}%, -${left}%)`,
+  };
+}
+
+const useStyles = makeStyles((theme) => ({
+  paper: {
+    position: "absolute",
+    width: 400,
+    backgroundColor: theme.palette.background.paper,
+    border: "2px solid #000",
+    boxShadow: theme.shadows[5],
+    padding: theme.spacing(2, 4, 3),
+  },
+}));
+
+const MartianModalContent = ({customer}) => {
+  const classes = useStyles();
+  const [modalStyle] = React.useState(getModalStyle);
+
+  return (
+    <>
+      <div style={modalStyle} className={classes.paper}>
+        <h2 id='simple-modal-title'>{customer.name}</h2>
+        <p id='simple-modal-description'>{customer.name}</p>
+      </div>
+    </>
+  );
+};
+
+export default MartianModalContent;
